refactor: extract HTTP SQL profiles into smugglr-core
Move Profile, AuthFormat, RequestFormat, and all profile constructors
from the http-sql plugin into smugglr-core. The plugin re-exports from
core. This enables the upcoming WASM fetch adapter to share the same
profile definitions without duplication.
