+++ plugins/smugglr-http-sql/Cargo.toml
@@ -13,6 +13,7 @@ name = "smugglr-http-sql"
 path = "src/main.rs"
 
 [dependencies]
+smugglr-core = { path = "../../crates/smugglr-core", default-features = false }
 smugglr-plugin-sdk = { path = "../../crates/smugglr-plugin-sdk" }
 reqwest = { version = "0.12", features = ["json"] }
 serde = { version = "1", features = ["derive"] }
